Add-ons: Publish HDMIPi add-on app and wiring schematic
HDMIPi add-on extends MuxPi with audio and video grabbing cpabilities.
It uses Lenkeng LKV373a HDMI Extender to turn HDMI output into MPEG
UDP network stream. This stream is displayed usign QVlc widget.
To provide ability to interact with the stream NanoPi is connected
to DUT via USB. It emulates a keyboard and a touch screen what allows
to collect input events from the user and pass them to a remote
target.

Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>